Create presenter for content_item
This will need to be split at some point in the not to distant future
into a case study presenter and a generic edition like presenter. But
for the moment it is in a generic content item presenter.

Also use components to make the page closer represent the direction it
is heading in.