First pass contract tests for content schemas
These tests will verify that government-frontend is able to render all the
currently available examples in govuk-content-schemas. We'll use these in the
build process of govuk-content-schemas to make sure changes to the schema do not
break the government-frontend application.

Note I've had to add a Spanish translation file because the schema examples
include a Spanish translation.