Display the "part of" metadata
I based this upon the code which currently does it in Whitehall:
https://github.com/alphagov/whitehall/blob/b1f827a2daa1df50cf69785af6827a621857e815/app/helpers/document_helper.rb#L205-L247

Note that the code in Whitehall has many more fields in it, but these aren't
currently sent to content-store. I think this may be based upon an assumption
that Case Studies only have the fields currently being sent to content-store.