Only run contract tests for specific formats
The government-frontend does not (yet) support the 'coming_soon' format.

Only attempt to run contract tests for formats which we currently
support.

As we evolve the contract testing approach, we may want to develop a
convention for frontends to declare which formats they support. For now
it seems reasonable to hard code it.

We filter examples by format rather than filename so that if new
examples are added for a supported format, we'll run those examples too.