Support contract testing in jenkins
Whenever a change is made to govuk-content-schemas we run the examples
against government-frontend to ensure that it still supports all of the
examples.

This will occur for branch builds as well.

This jenkins script provides a way to build the examples. It's
parameterised using `$SCHEMA_GIT_COMMIT`, so that it can be
tested against a particular branch or sha ref of the schemas.

https://trello.com/c/nmSJPc9Z/4-setup-ci-job-for-govuk-content-schemas