Support Coming Soon content items
Coming Soon content items are used for scheduled publishing. It is important in
scheduled publishing that all users see the new version as quickly as possible.
For content that is replacing existing content, this is easy: we simply wind in
the cache headers leading up to the publish time. For new content, this is
harder, because our stack doesn't respect the cache time for HTTP responses
with the 404 status. So instead, we return a 200 status and display a "Coming
Soon" page with the same cache header winding.