Add test for setting the correct Cache-Control header
It is important that we take the expiry time from content-store and apply it to
our responses. This is because content-store will vary the expiry time so that
scheduled publications are consistently visible at the time they are scheduled
for, rather than users having to wait for caches to clear.

This behaviour is always important for coming_soon items, but is needed for all
formats because updates can be scheduled to be published for them too.