Pass-in view_context to content item presenter
https://trello.com/c/VzKdQ6xm

we should prefer this method of using view helper methods over
mixing-in related modules in the presenter. asset URL helper
methods depend on `view_context` to access configuration and
request information, for example request protocol:

https://github.com/rails/rails/blob/v4.1.5/actionview/lib/action_view/helpers/asset_url_helper.rb#L194

as a result, `*_url` methods don't work in a presenter when
mixed-in, because they don't have access to `config` where it
looks for the asset host. without knowing asset_host it returns
a path instead of returning the full URL, and to make it worse,
does that silently.

this is done in preparation for introducing logic to create a
placeholder image URL in the next commit.