"Archived" content is now "Withdrawn"
This updates the display of previously "archived" content to be labelled as
"withdrawn", to be consistent with the retention policy guidance:
https://www.gov.uk/guidance/content-design/gov-uk-content-retention-and-withdrawal-archiving-policy.

It also updates the presenter code accordingly. Separate work will have to
happen to update the schema and migrate the existing content to use the term
"withdrawal" instead or "archive" in the content items themselves.