Show updated_at fields for service manual guides
Replace the placeholder value with an actual value from content-api. It
displays a human readable 'time ago' label and adds a title attribute to show
the actual time stamp on hover. The link is planned to reference a diff in
future, but it's not available yet. I'm leaving it as a link for styling
purposes while we work on it.

This commit also makes use of govuk content schema helpers to load an example
API response as a text fixture.