Display Service Manual Section pages
This is a part of work on Service Manual Section format.

Based on a spike [1] done in service-manual-publisher I've created a
govuk-content-schemas format [2] that represents Service Manual Section format.
Even though the publisher part is not finished, we believe the schema should
be somewhat stable and it's ok to build features around it.

[1] - https://github.com/alphagov/service-manual-publisher/tree/section_hierarchy_spike
[2] - https://github.com/alphagov/govuk-content-schemas/compare/service_manual_sections