Change presenter logic to use the 'topic' data format
it was decided to use the existing 'topic' format to represent service manual
topic pages, therefore the previously written code needed to be changed to use
this format. I also renames some methods and private classes, also moves them
to their own name-spaced files.

The code has been made more robust to deal with the fact that we allow
unpublished content items to be referenced in topic groups. If the content item
is not published, it will be removed from the "links -> linked items" hash and
the code will simply omit the referenced item. If all the items in a group are
omitted, then the whole group will be ignored too.