Add topic and org metadata to Stats Announcements
Originally in WH these were labelled as 'Organisations' and 'Topics',
but elsewhere in WH, in government-frontend, and in the metadata
component, we use 'From' and 'Part Of'.

Make these consistent, rather than adding a special case, but to be
discussed with a product manager.

Bottom margin on the metadata block was required for spacing, so
we converted it to a mixin, to avoid duplicating it with Case Study.