Add release date to Statistics Announcement
This introduced a slight rendering change from the original
Whitehall version, where the order of the metadata changes slightly,
with release date moving to the bottom, after org/topic.

We talked to Jen Allum who 👌 'd the change, as the ordering is
enforced by the metadata component and changing it isn't easy at
the moment. We may revisit this later.