Add national statistics logo logic
Some statistics announcements are national statistics, and legally
have to be marked as such, with the National Statistics logo.

The logo top position is offset to match the top margin of the
title component - this is not a good long term solution, and the top
margin of the title component can change independently to this repo.
This approach gets it working for now, to move through migration, but
we're going to revisit it before issuing a Pull Request.