Add format translation for statistics announcement
I tried running `rake translation:regenerate` to update the non-en
locales but encountered errors:
```
No pluralization forms defined for locale 'dr'. This probably means that the rails-18n gem does not provide a definition of the plural forms for this locale, you may need to define them yourself.
```

Adding just the english translation for now, as
  a) stats announcements aren't translated
  b) the rake task still has the same error without this change