Use GOV.UK Govspeak component for rendering SM guide body
Use the component because it has support for automatic youtube video link
embedding. It modifies the styling of text slightly, which is not ideal, but we
decided it's better than overwriting the components styles.

The long term plan is to list the discrepancies and update the component with
improvements that have been discovered through Service Manual user research.