Fetch content owner info from the links hash
As a part of an effort to reference everything using content_ids allow
Service manual guides to use information about the service manual guide content
owner from the links hash (see [1]) but fall back to the old way when the links
hash is not present.

This will allow us to migrate to the new format without breaking anything.

The fallback code should be removed once the migration process is complete.

[1] - https://github.com/alphagov/govuk-content-schemas/pull/219