Add breadcrumbs to Service manual topics
Breadcrumbs for Service manual topics only have the root of /service-manual and
itself as a text (rather than a link).

If we start nesting topics this will need to become smarter.

Also remove the duplicate breadcrumb Sass file, this is no longer required now
that the component is being used.