Refactor stats announcement presenter tests
Use content schema examples, rather than stubbing fields explicitly
in the test.

This more closely couples the presenter and the schemas repo and
changes to the later are more likely to cause test failures with
the former. But this is probably a good thing, as we're treating
schema changes as the authority in our system, if they or the
examples change we should have to update apps.

Also, this approach enforces that schema examples cover at least
enough cases to support testing, making it less likely to have
situations where a schema only has an example for the "happy case".
This also has the happy side-effect of ensuring frontends will
have more examples for manual testing against