Make title integration tests more specific
Previosly a simple `page.has_text?` assert would pass if the title
text appeared _anywhere_ on the page, not just in the title. This
could lead to false positive tests, where the title test passes
because the title text is elsewhere in the page copy (eg, repeated
in a description), but broken for the title component itself.

This change adds a scoped assert that ensures the text appears within
the title component, and specifically in the title argument passed to
the component.

This futher prevents false-positive passes, in a situation where the
asserted value may appear in another part of a component. This isn't
likely in title, but more likely in things like metadata, or document
footer, where there are many values.