Remove more duplication in format integration tests
I don't think we gain anything by duplicating the titles in the test,
as long as we're asserting the integration against the value expected
from the stubbed content item.

This might be a bit contentious, as we've gone a slightly different
direction with the presenter unit tests, but worth debating I think.