Refactor metadata component integration tests
Add a helper to reduce duplication, improve readability and increase
specificifity of test. Previously we only asserted that the right
labels and values were in the metadata component args, not that
they were paired together.

This helper specifically asserts that label and value are matched,
which I think reads more cleanly in the tests themselves.