Support Statistics announcements with changed date
This applies when a 'confirmed' announcement has it's release date
edited, which should not normally happen. When it does it must be
disclosed, and the reason explained.

In WH this disclosure is done as a separate headed section, with
a second 'metadata'-style block to format the date and reason.

Long term this may not be the right pattern for the page, and may
be consolidated into the main metadata block, or otherwise be
made more consistent with other document formats.

To get the feature migrated in the short term i've included a second
call to the metadata component, which causes some complexity in the
integration test, where there is now two metadata components when
asserting. In order to avoid the conflict, i've scoped the metadata
asserts for the `release_date_changed` example only, as the others
don't have multiple metadata blocks.

This does add a slightly spurious class to distinguish the first/
main metadata block, but it feels better than adding a lot of
complexity to the tests or test helpers. If this is a pattern we
encounter in more formats in the future, then we should look at
adding support to the test helpers for cases like this.