Don't collapse empty column of contents
When no contents are shown the grid wrapper is empty. Because it has no
height it collapses and the content shifts to the left.

To keep the design consistent with what’s currently live, introduce an
offset class which is used only when there is no sidebar.

Grid helpers like this are best used in the front-end toolkit, but
while this is a one-off it makes sense to put it here.