Build a breadcrumb from the links->parent array for topics.
Our existing #breadcrumbs method was hardcoded to only work for the service manual. Instead we build the breadcrumbs from the links->parent array so that the same presenter can be used for topics in other contexts.

A breadcrumb that only has one element, just the current page, isn't useful as it will be the same as the page title. Therefore if there aren't any parents supplied then we consider the breadcrumbs empty.