Reimplement content owners for service manual guides
A bit of to-ing and fro-ing here but this basically the same functionality reimplemented. Content owners can exist in the details or the links but with a preference for the links.

Once all guides have been republished we can remove the code that is trying to find the content owners in the details.