Add missing pluralization rules
This fixes broken rake tasks relating to translations (e.g. rake translation:regenerate)
and should ensure proper rendering of plurals for languages not included in rails-i18n.

There is a non ISO 639-1 code 'dr' which has been used for the
Dari (Afghan Persian) language. This matches the locale used in Whitehall but
should probably be 'prs' as per the ISO 639-3 standard.

All of the additional rules have been set to one/other to match the
available translations in the respective locale files.