Fix brand for executive office orgs in html pubs.
This will pass through the correct brand that the `organisation_logo`
component in `static` expects in order to display the `executive-office`
organisation correctly in HTML publications. More precisely, it will
trigger this class and its effects:
https://github.com/alphagov/static/blob/500d2fafc7280ba5cc295b28b03c1e85fdf3e95a/app/assets/stylesheets/govuk-component/_organisation-logo.scss#L45.

Relevant Trello ticket:
https://trello.com/c/cJGrtnc1/286-5-html-publications-migration-front-end-work-large

This is a hack that should be removed once organisations are migrated to
the content store.