Use single govspeak component
When using separate components the vertical rhythm of content could
break. We apply special rules to the first element in govspeak, eg
removing headings from the first H2.

The second govspeak component would remove the margin from the “Contact
details” heading, so its content would appear closer to the body than
it should.

Rather than wrap that content and give it some margin, concatenate the
two bodies and pass to a single component.