Run full suite of tests when contract testing on CI
Currently this app only runs a specific set of tests when triggered by
a new PR on govuk-content-schemas. This tripped us up today because
some tests that have been excluded are using the content schema
examples in the tests
(https://github.com/alphagov/government-frontend/pull/147).

The tests for this repo are really fast, so it’s not a problem to run
the entire test suite for each change on govuk-content-schemas. This
will prevents test runs to pass by accident and makes the build process
easier to understand.