Add meta descriptions to content items
The meta description tag is defined in the parent layout, but none of the content items currently use it. This commit adds meta description tags to all content items, the contents of which are set to the `description` associated with the content. This adds descriptions to external search engine results when the pages are re-indexed.