Fix breadcrumb presenter
This fixes a bug in the breadcrumbs presenter.

The presenter is tested with an example from govuk-content-schemas. But
the example contains an error: it uses a structure like `links ->
parent -> parent`. This is not how the publishing-api expands the
links, it nests it like `links -> parent -> links -> parent`.

However, because the liberal way the frontend `links` hash is
validated, the example was considered valid. This is tackled in
https://github.com/alphagov/govuk-content-schemas/pull/389.