Add ways to respond feature
Only displays when open and when a respond online link, email or postal
address have been provided. This matches existing logic. If an
attachment is uploaded but no email or postal address are provided to
return the completed copy it will not display. This is how it is in
Whitehall but could be improved.

Port of existing copy which isn’t great. Fixes the problem of "Respond
online" looking like a heading, instead using the call to action
pattern. Described here:
https://trello.com/c/pEccXKHV/517-issue-with-ways-to-respond-section-in-
consultations-medium

This highlights a problem with how we render content: this HTML is
crafted from values provided by users, rather than govspeak – yet it is
still content and must be rendered as content using the same styles.
For now this means passing that HTML to the govspeak content for
rendering.

An alternative would be some free text the user provides in markdown.
This would make the format more flexible but less uniform.