Fix tests for presenter#format
The content item attribute `format` is deprecated.

https://github.com/alphagov/govuk-content-schemas/pull/444 attempts to
remove `format` from the frontend, but the tests for this application
fail without `format` in the examples. This is because the tests are
testing kind of the wrong thing.

For example:

```
assert_equal schema_item['format'], presented_item.format
```

Checks the string in `presented_item.format` which actually points to
`schema_item['schema_name']`. Because they're `format` and
`schema_name` are identical in the examples this worked. Now that we've
removed `format` it doesn't anymore.