Improve display of share buttons on rtl content
When text-direction is set to right-to-left, the inline-block icons get
split weirdly, creating one link that wraps another.

On rtl set the container to inline-block, this treats the whole block
as a single run of text, rather than two. That prevents the link from
being split.

Also clean up the margins