Show breadcrumbs consistently
Currently we don't show the breadcrumbs if the format traditionally
doesn't have them.

On other pages on GOV.UK we've introduced the breadcrumbs, whether or
not they're set. For example, https://www.gov.uk/aaib-reports just has
a "Home" breadcrumb, because
https://www.gov.uk/aaib-reports/aaib-investigation-to-airbus-a319-111-g-
ezfj has a breadcrumb to that page.

This adds consistent breadcrumbs to all pages. This means that we'll
show empty "Home" breadcrumbs for fatality notices, statistics
announcements, statistical data sets, "take part" pages and groups.