Make breadcrumbs nice for HTML publications
https://github.com/alphagov/government-frontend/pull/238 introduced
breadcrumbs on all pages. Because HTML publications weren't in the list
of migrated formats in the readme (fixed in
https://github.com/alphagov/government-frontend/pull/239) I hadn't
checked that. @fofr helpfully pointed me to that.

Because HTML publications have its publication in the links as
`parent`, the breadcrumb magically shows the parent publication. This
means we can remove the link on the right. Because the breadcrumb takes
care of its own height, we no longer need to include the
responsive-top-margin mixin.