Return 406 when unacceptable format requested
Previously, calling a format without an atom feed via:
/government/consultations/postgraduate-doctoral-loans.atom

Would have 404’d. Now there is atom support this retrieves a content
item, but throws a “MissingTemplate” error when it comes to rendering.

Catch this error and return a “not acceptable” 406 response. eg The
target resource does not have a current representation that would be
acceptable to the user agent