Fix #181: Highlight nav correctly for news articles
Current nav is based on part of the path, which works for common cases
like 'publications', but some don't follow the convention, like news
articles (with 'government/news/slug') that sit under 'announcements'.

Add a hash to map paths that don't follow the convention.

Add tests to for news articles getting the right active nav, and
another to a format that does follow the convention, to ensure that
still works.

I'm not sure if there are formats other than news articles that need
a mapping, but they'd be easy to add now.