Move webchat to a legacy module folder
webchat is a legacy module that is not written in the
standard GOV.UK module pattern.

This module is going to be looked at being rewritten in firebreak.
while just calling it legacy and been done with it is not ideal,
we are looking at it in short order.

Also add contacts to readme and revert the change to application.js
back to just include the modules directory