Remove redundant stylesheet link
A snippet of code was used from this website:
https://makandracards.com/makandra/11133-how-to-test-print-stylesheets-with-cucumber-and-capybara
However, this snippet referenced a stylesheet called screen.css that was
never implemented on our end.

Since we never created the stylesheet, but we referenced it, all HTML
publications have been trying and failing to fetch this stylesheet,
resulting in a 404.

We don’t actually need screen.css, so this Pull Request removes the
reference to it.