Fallback to provided value if not found in allowed list
When content has a facet value that is not in the distinct set of
allowed values, rather than falling over and giving the user a 500,
instead render the value as it has been provided. Send a notification
to Errbit that there is an invalid value being presented.

Matches existing specialist-frontend behaviour.

Follows principles of:
1. Frontend rendering as best it can whatever has been provided by
content-store
2. Shouting (at least a little), when that content is invalid