This takes the answers code and makes it more generic so we can render the help page
I think maybe these two content types could be merged into one
to avoid duplicate code in the front end. With the css i have just followed
the current pattern of repetition, until we tackle it more wholesale.

i changed the gsub's for squish in the govspeak test helper, as it seems more
robust, and a content item was causing leading space problems.

It also adds the last updated block and gets rid of a rescue and reraise
that masked real errors making it hard to debug.