Send Airbrake notification when no parts found
A guide without parts is technically valid according to the schema. The
Publisher also currently allows it, although no guides have no parts.

We expect this to never be the case as a guide without parts shouldn’t
ever be approved and published. However, it could be true of the draft
content – ie a new content item without any parts _yet_.

Protect against no parts and send Airbrake notification if detected.