Add Wraith task for checking per format off the search api
It is now possible to run a rake task and see the top 10 per format from the search api,
using the rake command.

i also moved wraith gemfile out and moved it to the base directory. It was a total pain
on a day to day basis to have to change in and out of the directories, and this also makes
this script workable, without doing weird things like changing the directory on the user.