Show old related links for selected Guides
We are in the process of releasing the new navigation for Education
content. One item that needs to be solved from a product perspective is
how to deal with curated related links. Until that's solved, some
content items still need to show the old related links instead of the
new navigation for users in the B group of our A/B test.

These overrides existed in the frontend application. Since guides
were ported to government-frontend, we also need to port the override.