Add WorldLocationBasePath class
The WorldLocation format is about to become base_path-less as part of
the work to transition worldwide content to a taxonomy based structure.

Until we have a more permanent solution that associates the
world location to the appropriate taxon (which can then be used to
generate the base_path) we need a way of generating the correct
base_path for content that links to world locations.

This commit adds a class that will be used for this. Mostly, generating
the slug is simply a case of calling `parameterize` on the world
location title but there are a few exceptions to this which are also
handled by the `WorldLocationBasePath` class.