Fix WLNA metadata test
This test expects an incorrect base_path from the schema examples. The
examples are fixed but schemas can't pass CI currently as this test
fails now the base_path has been removed and is being inferred from the
world location title.

This commit will be updated with the correct test once schemas are
deployed.