Remove related links' overrides
We didn't support curated related links in the new navigation for some
time. That meant we had to show the old related links for some pieces of
content, due to the importance of the related links.

Now that we support it and the responsible departments agree, we are
showing the curated related links in the new navigation.

This means we don't need the manual overrides anymore. This commit
removes them so that the new navigation is presented.

Trello: https://trello.com/c/WrLDSJYm/26-deploy-new-curated-related-links-for-student-loans-content