Rename Parts to ContentParts
We do this to avoid collision with the Parts module in the multipart-post
gem that we're about to introduce when we replace airbrake with sentry.
This gem is a dependency of faraday which is introduced as a dependency of
sentry itself.  The Parts module from multipart-post is loaded before ours
and so when we reference Parts we get that one, not our one, and this
ultimately means our presenters don't have the correct methods and the app
breaks.

The name chosen is equally generic and could cause a problem in the future
but it matches the names of other similar modules.  In a future PR we'll
introduce a namespace for these modules to avoid the issue entirely as
it's likely we'd have the same problem with others (like Body or Metadata)
which are generically named.