Fix auto loading for Services
If we move `/lib/services.rb` to `/app/lib/services.rb` then it will be
available via rails autoloading by default because it lives in a
subfolder of `/app`.  When it lives in `/lib` we have to either require
it explicitly when we want to use it, or add `/lib` to the auto load
or eager load paths.  As `/lib` is already configured to be in those
paths but we're still getting load errors (see:
https://sentry.io/govuk/app-government-frontend/issues/355340514/) it
seems we've not set that up properly and moving it into `/app` will be
the safest option.