Introduce content_item namespace for presenter mixins
We do this because in #492 we introduced a new gem that had a `Parts`
module.  This clashed with our `Parts` module resulting in it not being
loaded, and our presenters not working.  So far we've only encountered
this problem once, and only with the `Parts` module, but it seems likely
to occur again as many of our mixins have generic names like `Body` or
`Metadata`.  In the previous PR we renamed `Parts` to `ContentParts` to
avoid the clash, but we discussed an alternative of introducing a
namespace for our modules to give them names that are much more likely to
be globally unique. That is what we do here, with the namespace
`ContentItem`.