Don't wrap numbers without a period
We've had a ticket come in where someone was writing dates in the
headings (e.g. 2 April 2017) and this was being parsed as a
numbered list. Since the guidance for govspeak is to include the
period after numbers you want to be recognised as numbers, I don't
think it makes sense for numbers without a period to be accepted.