Remove education AB test from government-frontend
Disabled the new breadcrumb and taxonomy related links. Essentially, rollback the B variant and always show A instead.

We will revisit this later and add extra logic to govuk_navigation helpers to only show the B variant on pages tagged to the new taxonomy. This PR strips away the AB logic to make it easier to do that work.