Add controller/integration testing
This commit adds testing for `answer` and `guide` pages, where it should
show the TaskList sidebar component.

The reason for mixing controller testing with integration is that there
seems to be an incompatibility with the `govuk_ab_testing` gem and
`Test::Unit` where its erring with the following:

```
NoMethodError: undefined method `header' for
Did you mean?  headers
               headers=`
```

In order to move forward I've decided to do the integration testing in
the controller testing.