Disable lint checking in imported css files
The css we are using for the A/B test has been imported from the
new "govuk-frontend". 

Rather than fixing the lint errors in two places, and as this 
test is temporary, we are moving the the css into a "vendor" folder,
as folders named "vendor" are automatically ignored by the linter.