Update content on forgot password page
The Verify link goes to the hub start page. This is because I am
not sure if it is possible to link directly to the middle of the
Verify flow without causing timeouts.

The Analytics team in Verify have asked that we add a query string
when linking to Verify from the B variant, so I have added
`?SelfAssessmentSigninTestVariant=B` to the end of the url.