Test taxonomy logic on pages not part of content nav test
Guides and detailed guides form part of the upcoming multivariate
content navigation test. The criteria for that test is “tagged to a
taxon” and a restricted guidance subset.

* Switch to document collection to test logic and avoid falling into
test (document collection is classified as guidance for the taxonomy
but isn’t in the stricter set used for content nav test)