Change the POST route for choose-sign-in to match the GET
The original POST route for the form on the choose-sign-in page hasn't
been registered in the environments. This means that when we tested in any
other test environment apart from development, the POST route returned
a 404. To get around this we have changed the route of the POST to match
the GET route which has already been registered.