Only state variant if content navigation test applies
The helper methods for checking the current variant (eg
`universal_navigation_without_nav?`) were not checking whether the AB
test applied or not.

If the CDN set the response header and this method was used on a page
that doesn’t qualify for the test, then it would give a false positive.

* Add test to confirm expected behaviour
* Update helpers to check AB test applies before checking variant