Add a guard to the SelfAssessment AB analytics tag
Add a guard to the analytics meta tag so it only shows on
pages relevant to SelfAssessmentSignTest A/B test and prevents
and possible clashes with other A/B tests that are running.

The guard re-uses an existing method in the controller, and
declares it as a "helper" method so that it can be used by
the application.html.erb view.

The guard needs to know about the content item to check the
value of the base path, so the tests have been refactored
to set up the a/b test variant after the content_item has
been created.