Move self assessment tests into their own file
The content_items_controller_test file was becoming cluttered
with tests that relate to the self assessment A/B test.

Moving the tests to their own file will make them easier to 
remove when the A/B test finishes. The `path_for` method has 
been duplicated rather than being moved to a helper class for
the same reason.