Create a new controller for SelfAssessment pages
Refactor the content_items controller and move everything that
is related to the new SelfAssessment signing pages to their own
controller. This change de-clutters the content items
controller, and will make it easier to remove all of the code
related to the Self Assessment A/B test when the test ends.

As the methods in the new self_assessment_ab_test controller
also need to be able to call `ab_test_applies?` in the Traffic
Signs Summary A/B test, I've moved that code to the Application
controller.

Tests have been updated.

Tech debt: I have duplicated two method:
1. the `path_for` method in the test files
2. the `content_item_path` in the Self Assessment concern.

I will move the `content_item_path` to the application controller in a
later commit.