Don't show new nav for world wide mainstream pages
We now determine which navigation to show according to a complex
decision tree. When we rolled it out we thought that it would be good
to always show the new navigation if the content is tagged to the world
wide taxonomy, because it would be closest to what the B-variant was.

This had (for me) unintended effects of showing the world wide
navigation on some mainstream pages, because they're tagged to the
"generic" taxons in the world wide taxonomy.

The worst example at the moment is "Set up a business", which is tagged
to the generic "Trade and invest" taxon, and therefore loses its useful
mainstream navigation (including related links).