Update taxonomy test to stub rummager request
This commit prepares the tests for the upgrade to GovukNavigationHelpers v7.0. Realistically, taxon content items always have content_ids. When calling `GovukNavigationHelpers::NavigationHelper.taxonomy_sidebar` for a content item with a taxon that has a content_id set, a request is made to Rummager for related content. Therefore we need to stub the request to make this test pass.