Update NavigationType to take presenter as param
We already have a method for retrieving taxons (`ContentItemPresenter#tagged_to_taxon?`), so we remove the code duplication by passing in the ContentItemPresenter object instead of the content store response hash. This prepares the code for switching out the logic for getting taxons from the content item in the next commit.