Add a homepage for developers
This adds a page at the root of the site that will be visible in
development and on Heroku review apps.

It has the following things that would be useful in development:

- Links to the component guide
- The URLs of the content store, rummager, static, so it's clear which
version we are using
- Link to the canonical example of a template (the one named after the
schema)
- Link to a random version of the template
- List of links to pages rendered by government-frontend. I've reused
the wraith config for this, it seemed like a nice representative list
of different pages.

https://trello.com/c/IoHKyhoX