Use NavigationHelper content item to fetch taxons
Since this PR[1] we aim to only show taxons that are descendent from
the three published root taxons (education, parenting and world). This
involved adding some code that recursively walks up the parent taxons
and checks if the root taxon is whitelisted.
However in `ContentItemPresenter#tagged_to_a_taxon?` we were only
looking at the taxons in the content store payload, so effectively
skipping the logic that whitelisted the root taxons. This commit
updates the class so that we use the code from
`GovukNavigationHelpers::ContentItem#parent_taxons`[2] instead.

[1] https://github.com/alphagov/govuk_navigation_helpers/pull/82
[2] https://github.com/alphagov/govuk_navigation_helpers/blob/02acca91a280a5e5141bf6e5cfb2b455f955c8a9/lib/govuk_navigation_helpers/content_item.rb#L38-L44