Split hint text from label and link them with `aria-describedby`
With long labels, by default this component visually makes the most important content
distinguishable, by using bold and color. We can't rely on this for users with disabilities.

If someone has difficulty seeing our pages, they might use assistive technologies such as screen readers.
In this case, if all our text is within our labels (even if visually they're distinct) that text will be announced
in it's entireity.

By splitting the hint text up and having it alongside our labels, we can then associate them with `aria-describedby`.

This results in the main label being announced, then an indication that this is a radio component and only after that is the additional text
read out. So it means users of screenreaders can scan and skip options just as easy as people making choices with sight.

The classes have to be reorganised to achieve this, since as explained before, these elements now have to be completely separated.

See https://github.com/alphagov/govuk_elements/issues/574 for a prior discussion around this